+++ fontforge/parsettf.c
@@ -4123,7 +4123,7 @@ static void readttfpostnames(FILE *ttf,struct ttfinfo *info) {
 	    sprintf(buffer, "%.20s-%d", info->ordering, i );
 	else if ( info->map!=NULL && info->map->map[i]!=-1 )
 	    sprintf(buffer, "nounicode-%d-%d-%x", info->platform, info->specific,
-		    info->map->map[i] );
+		    (int) info->map->map[i] );
 	else
 	    sprintf( buffer, "glyph%d", i );
 	info->chars[i]->name = copy(buffer);
